src/requirements.txt
python-dotenv
argparse
requests
pysheds
tqdm
torchvision